requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
cerebras-cloud-sdk>=1.0.0

